test: add tests for KV bindings
